+++ tsconfig.json
@@ -27,7 +27,12 @@
     "src/common/load-plugins.js",
     "src/cli/index.js",
     "src/cli/expand-patterns.js",
-    "src/cli/util.js",
+    "src/cli/core.js",
+    "src/cli/context.js",
+    "src/cli/format.js",
+    "src/cli/option-map.js",
+    "src/cli/option.js",
+    "src/cli/usage.js",
     "src/cli/constant.js",
     "src/cli/prettier-internal.js",
     "src/languages.js",
